Creates separate unit test application for filter-related tests.
- Addresses issue #16
- The 16- and 32-bit FIR filter and 32-bit biquad filter tests have been removed from the vect_tests application and a new unit test application (filter_tests) was added.
- filter_tests does not add any new behavior (such as benchmarking performance, or interaction with python/numpy) at this time.
