+++ test/CMakeLists.txt
@@ -8,3 +8,4 @@ add_subdirectory( shared )
 add_subdirectory( vect_tests )
 add_subdirectory( bfp_tests  )
 add_subdirectory( fft_tests  )
+add_subdirectory( filter_tests  )
