+++ test/filter_tests/config.xscope
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!-- ======================================================= -->
+<!-- The 'ioMode' attribute on the xSCOPEconfig              -->
+<!-- element can take the following values:                  -->
+<!--   "none", "basic", "timed"                              -->
+<!--                                                         -->
+<!-- The 'type' attribute on Probe                           -->
+<!-- elements can take the following values:                 -->
+<!--   "STARTSTOP", "CONTINUOUS", "DISCRETE", "STATEMACHINE" -->
+<!--                                                         -->
+<!-- The 'datatype' attribute on Probe                       -->
+<!-- elements can take the following values:                 -->
+<!--   "NONE", "UINT", "INT", "FLOAT"                        -->
+<!-- ======================================================= -->
+
+<xSCOPEconfig ioMode="basic" enabled="true">
+
+    <!-- For example: -->
+    <!-- <Probe name="Probe Name" type="CONTINUOUS" datatype="UINT" units="Value" enabled="true"/> -->
+    <!-- From the target code, call: xscope_int(PROBE_NAME, value); -->
+    
+    <!--<Probe name="out_buffer_level"       type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!--<Probe name="GC_GAIN" type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/> -->   
+    <!-- <Probe name="out_buffer_level"       type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="samples_out"            type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="peak_association_time"  type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="start_bin"         type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="resort_time"       type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="peak_count"        type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="samples_out"       type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="frame_recv_time"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="frame_send_time"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="xcspe"       type="CONTINUOUS" datatype="INT" units="Value" enabled="false"/>  -->
+    <!-- <Probe name="gain"       type="CONTINUOUS" datatype="INT" units="Value" enabled="false"/>  -->
+    <!-- <Probe name="fit_count"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="timing_application_task" type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="timing_singlet_fit"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="timing_speaker_model"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="timing_fitter"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="timing_resynth"    type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+    <!-- <Probe name="timing_td_detection"    type="CONTINUOUS" datatype="INT" units="Value" enabled="false"/>  -->
+    <!-- <Probe name="timing_kde" type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>  -->
+</xSCOPEconfig>
