+++ test/filter_tests/src/filter/test_xs3_filter_fir_s32.c
@@ -11,7 +11,7 @@
 
 #include "../src/vect/vpu_helper.h"
 
-#include "../../tst_common.h"
+#include "../tst_common.h"
 
 #include "unity.h"
